+++ modules/Student_Billing/MassAssignPayments.php
@@ -4,22 +4,34 @@ if ( $_REQUEST['modfunc'] === 'save' )
 {
 	if (count($_REQUEST['student']) && AllowEdit())
 	{
-		//FJ fix SQL bug invalid amount
-		if (is_numeric($_REQUEST['amount']))
+		$date = RequestedDate(
+			$_REQUEST['year_date'],
+			$_REQUEST['month_date'],
+			$_REQUEST['day_date']
+		);
+
+		// FJ fix SQL bug invalid amount.
+		if ( is_numeric( $_REQUEST['amount'] ) )
 		{
-			foreach ( (array) $_REQUEST['student'] as $student_id => $yes)
+			if ( $date )
 			{
-				$sql = "INSERT INTO BILLING_PAYMENTS (ID,SYEAR,SCHOOL_ID,STUDENT_ID,PAYMENT_DATE,AMOUNT,COMMENTS)
-							values(".db_seq_nextval('BILLING_PAYMENTS_SEQ').",'".UserSyear()."','".UserSchool()."','".$student_id."','".DBDate()."','".preg_replace('/[^0-9.-]/','',$_REQUEST['amount'])."','".$_REQUEST['comments']."')";
-				DBQuery($sql);
+				foreach ( (array) $_REQUEST['student'] as $student_id => $yes)
+				{
+					$sql = "INSERT INTO BILLING_PAYMENTS (ID,SYEAR,SCHOOL_ID,STUDENT_ID,PAYMENT_DATE,AMOUNT,COMMENTS)
+								values(" . db_seq_nextval( 'BILLING_PAYMENTS_SEQ' ) . ",'" . UserSyear() . "','" . UserSchool() . "','" . $student_id . "','" . $date . "','" . preg_replace( '/[^0-9.-]/', '', $_REQUEST['amount'] ) . "','" . $_REQUEST['comments'] . "')";
+					DBQuery($sql);
+				}
+
+				$note[] = button('check') .'&nbsp;'._('That payment has been added to the selected students.');
 			}
-			$note[] = button('check') .'&nbsp;'._('That payment has been added to the selected students.');
+			else
+				$error[] = _( 'The date you entered is not valid' );
 		}
 		else
-			$error[] = _('Please enter a valid Amount.');
+			$error[] = _( 'Please enter a valid Amount.' );
 	}
 	else
-		$error[] = _('You must choose at least one student.');
+		$error[] = _( 'You must choose at least one student.' );
 
 	unset($_SESSION['_REQUEST_vars']['modfunc']);
 	unset($_REQUEST['modfunc']);
@@ -46,9 +58,20 @@ if ( ! $_REQUEST['modfunc'] )
 
 		echo '<table class="col1-align-right">';
 
-		echo '<tr><td>'._('Payment Amount').'</td><td><input type="text" name="amount" size="5" maxlength="10" required /></td></tr>';
+		echo '<tr><td>' . _( 'Payment Amount' ) . '</td>
+			<td><input type="text" name="amount" size="5" maxlength="10" required /></td></tr>';
 
-		echo '<tr><td>'._('Comment').'</td><td><input type="text" name="comments" /></td></tr>';
+		echo '<tr><td>' . _( 'Date' ) . '</td>
+			<td>' . DateInput(
+				DBDate(),
+				'date',
+				'',
+				false,
+				false
+			) . '</td></tr>';
+
+		echo '<tr><td>' . _( 'Comment' ) . '</td>
+			<td><input type="text" name="comments" /></td></tr>';
 
 		echo '</table>';